tests: make use of stream_helpers in append_break.cpp
