+++ tests/integrity/append_break.cpp
@@ -5,79 +5,11 @@
  * append_break.cpp -- pmemstream_append break - data integrity test
  */
 
+#include "stream_helpers.hpp"
 #include "unittest.hpp"
 
 #include <cstring>
 #include <vector>
-
-namespace
-{
-/* XXX: these helper methods are copy-pasted from "append.cpp" RC test.
- *	We need to make them usable for all (RC and non-RC) tests (use defines/macros for ASSERTs?). */
-
-/* append all data in the vector at the offset */
-void append(struct pmemstream *stream, struct pmemstream_region region, const std::vector<std::string> &data)
-{
-	for (const auto &e : data) {
-		auto ret = pmemstream_append(stream, region, nullptr, e.data(), e.size(), nullptr);
-		UT_ASSERTeq(ret, 0);
-	}
-}
-
-/* initialize stream with a single region and initial data (if given) */
-struct pmemstream_region init_stream_single_region(struct pmemstream *stream, size_t region_size,
-						   const std::vector<std::string> *data = nullptr)
-{
-	struct pmemstream_region new_region;
-	UT_ASSERTeq(pmemstream_region_allocate(stream, region_size, &new_region), 0);
-
-	struct pmemstream_entry_iterator *eiter;
-	UT_ASSERTeq(pmemstream_entry_iterator_new(&eiter, stream, new_region), 0);
-
-	/* Find out offset for the first entry in region */
-	struct pmemstream_entry entry;
-	UT_ASSERTeq(pmemstream_entry_iterator_next(eiter, NULL, &entry), -1);
-	pmemstream_entry_iterator_delete(&eiter);
-
-	append(stream, new_region, *data);
-
-	return new_region;
-}
-
-/* get first (and only) region */
-struct pmemstream_region get_first_region(struct pmemstream *stream)
-{
-	struct pmemstream_region_iterator *riter;
-	int ret = pmemstream_region_iterator_new(&riter, stream);
-	UT_ASSERTne(ret, -1);
-
-	struct pmemstream_region region;
-	ret = pmemstream_region_iterator_next(riter, &region);
-	UT_ASSERTne(ret, -1);
-	pmemstream_region_iterator_delete(&riter);
-
-	return region;
-}
-
-/* read all elements in a region */
-std::vector<std::string> get_elements_in_region(struct pmemstream *stream, struct pmemstream_region *region)
-{
-	std::vector<std::string> result;
-
-	struct pmemstream_entry_iterator *eiter;
-	UT_ASSERTeq(pmemstream_entry_iterator_new(&eiter, stream, *region), 0);
-
-	struct pmemstream_entry entry;
-	while (pmemstream_entry_iterator_next(eiter, NULL, &entry) == 0) {
-		auto data_ptr = reinterpret_cast<const char *>(pmemstream_entry_data(stream, entry));
-		result.emplace_back(data_ptr, pmemstream_entry_length(stream, entry));
-	}
-
-	pmemstream_entry_iterator_delete(&eiter);
-
-	return result;
-}
-} /* namespace */
 
 static void test(int argc, char *argv[])
 {
@@ -95,7 +27,7 @@ static void test(int argc, char *argv[])
 		/* append initial data to a new stream */
 
 		auto s = make_pmemstream(path, TEST_DEFAULT_BLOCK_SIZE, TEST_DEFAULT_STREAM_SIZE);
-		init_stream_single_region(s.get(), TEST_DEFAULT_REGION_SIZE, &init_data);
+		initialize_stream_single_region(s.get(), TEST_DEFAULT_REGION_SIZE, init_data);
 
 	} else if (argv[1][0] == 'b') {
 		/* break in the middle of an append */
@@ -116,7 +48,7 @@ static void test(int argc, char *argv[])
 		auto r = get_first_region(s.get());
 
 		/* read back data and count for the same output */
-		auto read_elements = get_elements_in_region(s.get(), &r);
+		auto read_elements = get_elements_in_region(s.get(), r);
 		/* While iterating over all entries, entry torn
 		 * in the previous append should be cleared now. */
 		auto cnt = read_elements.size();
